almost done

+++ BL/DroneInCharge.cs
@@ -10,7 +10,7 @@ namespace IBL.BO
    public class DroneInCharge
     {
         public string IdNumber { set; get; }
-        public double Battery { set; get; }
+        public int Battery { set; get; }
         public override string ToString()
         {
             return this.stringProperty();
